requirements.txt
aniso8601==7.0.0
anytree==2.8.0
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-admin-thumbnails==0.2.6
django-ckeditor==6.5.1
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-extensions==3.2.1
django-filter==22.1
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
drf-extra-fields==3.4.1
drf-yasg==1.21.4
flake8==6.0.0
flake8-typing-imports==1.14.0
googletrans==3.0.0
graphene==2.1.9
graphene-django==2.15.0
graphene-django-filter==0.6.4
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
packaging==23.0
phonenumbers==8.13.4
Pillow==9.4.0
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
rfc3986==1.5.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stringcase==1.2.0
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.14.1
